Move the `_start` function into `src/arch/*`.
All the other inline asm blocks are in `src/arch/*`, so for consistency
move the `_start` function definition into src/arch/* too.

Also move the log initialization function into its own file.

After this, src/lib.rs doesn't have any significant code in it, which
makes it easier to see how the crate is organized.
