+++ src/arch/aarch64.rs
@@ -1,4 +1,3 @@
-#[cfg(any(feature = "origin-thread", feature = "origin-signal"))]
 use core::arch::asm;
 #[cfg(feature = "origin-signal")]
 use linux_raw_sys::general::__NR_rt_sigreturn;
@@ -11,7 +10,35 @@ use {
     rustix::thread::RawPid,
 };
 
+/// The program entry point.
+///
+/// # Safety
+///
+/// This function must never be called explicitly. It is the first thing
+/// executed in the program, and it assumes that memory is laid out according
+/// to the operating system convention for starting a new program.
+#[cfg(all(feature = "origin-program", feature = "origin-start"))]
+#[naked]
+#[no_mangle]
+pub(super) unsafe extern "C" fn _start() -> ! {
+    // Jump to `entry`, passing it the initial stack pointer value as an
+    // argument, a null return address, a null frame pointer, and an aligned
+    // stack pointer. On many architectures, the incoming frame pointer is
+    // already null.
+    asm!(
+        "mov x0, sp",   // Pass the incoming `sp` as the arg to `entry`.
+        "mov x30, xzr", // Set the return address to zero.
+        "b {entry}",    // Jump to `entry`.
+        entry = sym super::program::entry,
+        options(noreturn),
+    )
+}
+
 /// A wrapper around the Linux `clone` system call.
+///
+/// This can't be implemented in `rustix` because the child starts executing at
+/// the same point as the parent and we need to use inline asm to have the
+/// child jump to our new-thread entrypoint.
 #[cfg(feature = "origin-thread")]
 #[inline]
 pub(super) unsafe fn clone(
