+++ src/arch/riscv64.rs
@@ -1,14 +1,47 @@
+#[cfg(any(
+    feature = "origin-thread",
+    all(feature = "origin-program", feature = "origin-start")
+))]
+use core::arch::asm;
 #[cfg(feature = "origin-thread")]
 use {
     alloc::boxed::Box,
     core::any::Any,
-    core::arch::asm,
     core::ffi::c_void,
     linux_raw_sys::general::{__NR_clone, __NR_exit, __NR_munmap},
     rustix::thread::RawPid,
 };
 
+/// The program entry point.
+///
+/// # Safety
+///
+/// This function must never be called explicitly. It is the first thing
+/// executed in the program, and it assumes that memory is laid out according
+/// to the operating system convention for starting a new program.
+#[cfg(all(feature = "origin-program", feature = "origin-start"))]
+#[naked]
+#[no_mangle]
+pub(super) unsafe extern "C" fn _start() -> ! {
+    // Jump to `entry`, passing it the initial stack pointer value as an
+    // argument, a null return address, a null frame pointer, and an aligned
+    // stack pointer. On many architectures, the incoming frame pointer is
+    // already null.
+    asm!(
+        "mv a0, sp",    // Pass the incoming `sp` as the arg to `entry`.
+        "mv ra, zero",  // Set the return address to zero.
+        "mv fp, zero",  // Set the frame address to zero.
+        "tail {entry}", // Jump to `entry`.
+        entry = sym super::program::entry,
+        options(noreturn),
+    )
+}
+
 /// A wrapper around the Linux `clone` system call.
+///
+/// This can't be implemented in `rustix` because the child starts executing at
+/// the same point as the parent and we need to use inline asm to have the
+/// child jump to our new-thread entrypoint.
 #[cfg(feature = "origin-thread")]
 #[inline]
 pub(super) unsafe fn clone(
